+++ components/usedelay.tsx
@@ -1,22 +1,30 @@
-import { useEffect, useState } from "react";
+import React, { CSSProperties, useEffect, useState } from "react";
 
-function useDelayUnmount(isMounted: boolean, delayTime: number) {
-    const [ shouldRender, setShouldRender ] = useState(false);
-
-    useEffect(() => {
-    let timeoutId: any;
-        if (isMounted && !shouldRender) {
-            setShouldRender(true);
-        }
-        else if(!isMounted && shouldRender) {
-            timeoutId = setTimeout(
-                () => setShouldRender(false), 
-                delayTime
-            );
-        }
-        return () => clearTimeout(timeoutId);
-    }, [isMounted, delayTime, shouldRender]);
-    return shouldRender;
+interface useMountAnimationProps {
+    render: JSX.Element,
+    fallback: JSX.Element,
 }
 
-export default useDelayUnmount;
+interface useMountAnimationReturn {
+    mount:boolean,
+    mountedStyle: CSSProperties,
+    unmountedStyle: CSSProperties,
+    ContentOrSkeleton: JSX.Element
+}
+
+const useMountAnimation = ({render, fallback}:useMountAnimationProps): useMountAnimationReturn => {
+    const [shouldRender, setShouldRender] = useState<boolean>(false);
+    const mountedStyle:CSSProperties = { opacity: 1, transition: "opacity 1500ms ease-in" };
+    const unmountedStyle:CSSProperties = {
+        opacity: 0.5,
+        transition: "opacity 1500ms ease-in",
+    };
+    useEffect(() => {
+        setShouldRender(true);
+    })
+    return {mount: shouldRender, mountedStyle: mountedStyle, unmountedStyle: unmountedStyle, ContentOrSkeleton: shouldRender ? <> {render} </> : <>{fallback}</>};
+}
+
+export default useMountAnimation;
+
+